+++ src/util.ts
@@ -48,3 +48,17 @@ export async function runNpm(
     args: string[], options?: ExecFileOptions): Promise<string|Buffer> {
   return promisify(execFile)(npmCmd, args, options).then(({stdout}) => stdout);
 }
+
+/**
+ * Promisified version of setTimeout.
+ */
+export const wait = (ms: number) =>
+    new Promise((resolve) => setTimeout(resolve, ms));
+
+/**
+ * A function that should never be called. But if it somehow is anyway, throw an
+ * exception with the given message.
+ */
+export function throwUnreachable(_unreachable: never, message: string): void {
+  throw new Error(message);
+}